+++ src/imageBase.ml
@@ -95,6 +95,8 @@ module Pixel = struct
     let black = (0,0,0,0xff)
 
     let white = (0xff,0xff,0xff,0xff)
+
+    let transparent = (0,0,0,0)
   end
 
   external yuv_of_rgb : rgb -> yuv = "caml_yuv_of_rgb"